Move code into separate package
